requirements.txt
flask
requests

torch
torchvision
transformers
accelerate
bitsandbytes
llava-torch
Pillow
